% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame input that record X variable.}

\item{cl}{Vector input of true class value of training data.}

\item{k_nn, }{Numeric input representing the number of neighbors.}

\item{k_cv, }{Numeric input representing the number of folds.}
}
\value{
a list includes \code{class}, a vector of the predicted class y-hat
for all observations, \code{cv_err}, a numeric of the cross-validation
misclassification error.
}
\description{
This function does the Cross Validation test for the kth nearest neighbor
predict model.
}
\examples{
penguins_data <- na.omit(my_penguins)
train <- penguins_data[, 3:6]
cl <- penguins_data$species
my_knn_cv(train, cl, 5, 5)

}
\keyword{prediction}
