% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Constructor}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object input shows the relationship between
independ and depend variable.}

\item{data}{Data frame input contains the variables in the model.}
}
\value{
A table includes \code{Estimate}, a column of Estimate Coefficients;
\code{Std. Error}, a column of Residual Standard Error; \code{t value}, a
column of \code{Estimate} divided by \code{Std.Error}; and \code{p value},
a column of P-values corresponding to \code{t value} in the T distribution.
}
\description{
This function does the linear regression depend on the input formula and data
and return a table as result.
}
\examples{
data(mtcars)
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{inference,}
\keyword{prediction}
