% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input of number of folds.}
}
\value{
a numeric of the cross-validation error value.
}
\description{
This function does the Cross Validation test for the random forest predict
model. This function is specifically using \code{bill_length_mm},
\code{bill_depth_mm}, and \code{flipper_length_mm} to predict
\code{body_mass_g} from the data \code{my_penguins}.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
