% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test Calculator}
\usage{
my_t.test(x, alternative = "two.sided", mu = 0)
}
\arguments{
\item{x}{Numeric vector input, containing the values we are going to test.}

\item{alternative}{String input specifying alternative hypothesis, defaults
to \code{"two.sided"}.}

\item{mu}{Numeric input, indicating the null hypothesis value of the mean,
defaults to \code{0}.}
}
\value{
A list includes \code{test_stat}, the numeric test statistic;
\code{df}, the degrees of freedom; \code{alternative}, the value of the
parameter alternative; and \code{p_val}, the numeric p-value.
}
\description{
This function does the T-test with the input value, return the result as a list.
}
\examples{
x <- rnorm(10, mean = 0, sd = 1)
my_t.test(x)
my_t.test(x, alternative = "less")
my_t.test(x, alternative = "greater", mu = 1)

}
\keyword{inference}
