% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model for data}
\usage{
my_lm(formula, da)
}
\arguments{
\item{formula}{model formula}

\item{da}{data set}
}
\value{
A list includes linear model and model summary.
}
\description{
Linear Model for data
}
\examples{
data(cars)
my_lm(speed~dist, data=cars)

}
\keyword{prediction}
