% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation for penguins data}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input of number of folds.}
}
\value{
a numeric of the cross-validation error value.
}
\description{
This function does the Cross Validation test for the random forest predict model.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
